Almost all strings and buffers are now of type char. This fixes more than half the warnings on gcc4, and makes sure we do not get signedness problems between platforms where char is defined with different signedness. Also set line-ending properites for some bison/flex files that did not have it set.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@563 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/framework/types.h
@@ -103,8 +103,8 @@ typedef struct {
 
 typedef struct
 {
-  long  x;
-  long  y;
+  int  x;
+  int  y;
 } POINT;
 
 #endif /* !WIN32 */
